Move enums to constants
