+++ src/app/utils/constants.js
@@ -47,3 +47,11 @@ export const SYSTEM_MODE = {
   ON: 3,
   OFF: 4
 }
+
+export const VIEWS = {
+  CONNECTING: "CONNECTING",
+  METRICS: "METRICS",
+  AMPERAGE_SELECTOR: "AMPERAGE_SELECTOR",
+  REMOTE_CONSOLE: "REMOTE_CONSOLE",
+  MQTT_UNAVAILABLE: "MQTT_UNAVAILABLE"
+}
