--- src/config/enums.js
@@ -1,7 +1,0 @@
-export const VIEWS = {
-  CONNECTING: "CONNECTING",
-  METRICS: "METRICS",
-  AMPERAGE_SELECTOR: "AMPERAGE_SELECTOR",
-  REMOTE_CONSOLE: "REMOTE_CONSOLE",
-  MQTT_UNAVAILABLE: "MQTT_UNAVAILABLE"
-}
